Dropped redundant whitespace in the "skills" popup Done cleanup of the component
