+++ components/page/member-details/MemberDetailHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './MemberDetailHeader';
